add doc, rename redis store, bump deps, release 2.0.0
