+++ packages/serve1090/.env-template
@@ -1,2 +1,5 @@
-PORT=
+PORT=80
 SECRET=
+KEYDB_HOST=127.0.0.1
+KEYDB_PORT=6379
+KEYDB_PASSWORD=skyhawk
