+++ packages/serve1090/package.json
@@ -1,6 +1,6 @@
 {
   "name": "serve1090",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "private": true,
   "description": "WebSocket server for receiving dump1090 data.",
   "repository": "https://github.com/robertsteilberg/serve1090/",
@@ -8,12 +8,19 @@
   "author": "Robert Steilberg <rsteilberg@gmail.com>",
   "scripts": {
     "start": "yarn && node index.js",
-    "start:pm2": "pm2 start index.js",
+    "serve": "docker-compose up -d --build",
+    "serve:debug": "yarn redis-restart && yarn debug",
     "debug": "yarn && nodemon --inspect index.js --ignore logs",
+    "redis": "docker run -v $PWD/keydb.conf:/etc/keydb/keydb.conf --publish=6379:6379 --name=serve1090_keydb -d eqalpha/keydb:latest",
+    "redis-cli": "keydb-cli -a skyhawk",
+    "redis-stop": "docker stop serve1090_keydb",
+    "redis-remove": "docker rm serve1090_keydb",
+    "redis-restart": "yarn redis-stop && yarn redis-remove && yarn redis",
+    "start-pm2": "pm2 start index.js",
     "purge": "rm -rf logs/*"
   },
   "dependencies": {
-    "@hapi/joi": "^17.1.1",
+    "joi": "^17.2.0",
     "@turf/helpers": "^6.1.4",
     "@turf/boolean-point-in-polygon": "^6.0.1",
     "bree": "^1.1.21",
