+++ packages/serve1090/src/stores/aircraft-store.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const { store: logger } = require('../lib/logger');
+const logger = require('../lib/logger').get('store');
 const AIRCRAFT_SCHEMA = require('./schemas');
 const {
   secondsToMillis,
@@ -13,96 +13,12 @@ const {
 // maximum age of new data that will be accepted into the store
 const MAX_DATA_AGE = 10000;
 
-module.exports = {
-  init,
-  setNewData,
-  getRawAircraft,
-  getValidAircraft,
-  getExcludedAircraft,
-  shutdown
-};
-
-// INIT
-
-function init () {
-  if (!this.initialized) {
-    // set up data maps
-    this.currentRawAircraft = _createStore('raw');
-    this.currentValidAircraft = _createStore('valid');
-    // init jobs
-    this.jobs = [
-      setInterval(() => this.currentValidAircraft = _purgeAircraft(this.currentValidAircraft, 10000), 1000),
-      setInterval(() => this.currentRawAircraft = _purgeAircraft(this.currentRawAircraft, 60000), 10000),
-      setInterval(() => _emptyPurgeList(this.currentRawAircraft, this.currentValidAircraft), 1800000)
-    ];
-  } else {
-    shutdown();
-    init();
-    logger.info('store re-initialized');
-  }
-  this.initialized = true;
-  logger.info('initialize store');
-}
-
-// JOBS
-
-/**
- * Purge aircraft out of the store that are stale
- *
- * @param store the serve1090 aircraft store to purge
- * @param maxAge maximum allowed age of the aircraft (where age begins when serve1090
- *        receives the aircraft or an update to it)
- * @returns new serve1090 aircraft store without stale aircraft
- * @private
- */
-function _purgeAircraft (store, maxAge) {
-  const result = _.cloneDeep(store);
-  let newSize = Object.keys(store.aircraft).length;
-  result.aircraft = _.pickBy(store.aircraft, function (aircraft) {
-    const age = Date.now() - aircraft.updated;
-    const tooOld = age > maxAge;
-    if (tooOld) {
-      result.purged.add(aircraft.hex);
-      newSize--;
-      logger.info({
-        message: 'purge aircraft',
-        store: store.name,
-        hex: aircraft.hex,
-        age,
-        max: maxAge,
-        numPurged: result.purged.size,
-        newSize
-      });
-    }
-    return !tooOld;
-  });
-  return result;
-}
-
-/**
- * Empty the purge list, since the same aircraft will be seen over time
- * @param store the store containing the purge list to clear
- * @private
- */
-function _emptyPurgeList (...store) {
-  store.forEach(s => {
-    if (s.purged.size) {
-      s.purged.clear();
-      logger.info({
-        message: 'empty purge list',
-        store: s.name
-      });
-    }
-  });
-}
-
-// SHUTDOWN
-
-function shutdown () {
-  this.jobs.map(clearInterval);
-  this.initialized = false;
-  logger.info('shutdown store');
-}
+module.exports = (client) => ({
+  setNewData: setNewData.bind(null, client),
+  getAllAircraft: getAllAircraft.bind(null, client),
+  getValidAircraft: getValidAircraft.bind(null, client),
+  getInvalidAircraft: getInvalidAircraft.bind(null, client)
+});
 
 // CREATE
 
@@ -112,8 +28,7 @@ function shutdown () {
  * @param data raw JSON object generated by dump109 with data.now in seconds since epoch
  * @returns {boolean} true if the data was accepted, false if it was rejected
  */
-function setNewData (data) {
-  _checkIfInitialized(this.initialized);
+async function setNewData (client, data) {
   const clientNowMillis = secondsToMillis(data.now);
   const now = Date.now();
   const age = now - clientNowMillis;
@@ -125,73 +40,26 @@ function setNewData (data) {
     });
     throw new StaleDataError(millisToSeconds(age));
   }
-  logger.info({
-    message: 'accept dump1090 data',
-    messages: data.messages,
-    clientTimestamp: new Date(clientNowMillis).toISOString()
-  });
+  // logger.info({
+  //   message: 'accept dump1090 data',
+  //   messages: data.messages,
+  //   clientTimestamp: new Date(clientNowMillis).toISOString()
+  // });
   // first, update and filter the data
-  const newAircraftMap = _mapifyAircraftArray(data.aircraft.map(_setUpdated));
-  const { validatedMap, errors } = _validateAircraftMap(newAircraftMap);
-  // set errors onto newAircraftMap so they can be returned with raw/excluded
-  Object.entries(errors).forEach(([hex, error]) => {
-    newAircraftMap[hex].error = error;
-  });
-  // then merge it with existing data stores
-  this.currentRawAircraft = _mergeIntoStore(this.currentRawAircraft, newAircraftMap, this.client);
-  this.currentValidAircraft = _mergeIntoStore(this.currentValidAircraft, validatedMap, this.client);
-
-  return true;
+  const newAircraftMap = mapifyAircraftArray(data.aircraft.map(setUpdated));
+  await validateAndWrite(client, newAircraftMap);
 }
 
 /**
- * Merge a mapified dump1090 JSON object into an existing serve1090 aircraft
- * data store
  *
- * @param store existing serve1090 store of aircraft
- * @param newMap unmerged serve1090 map of aircraft | { hex: aircraft }
- * @returns newly merged aircraft data store | { hex: aircraft }
- * @private
+ * @param aircraftArray array of aircraft from dump1090
+ * @returns { hex: aircraft } hex mapped to aircraft object
  */
-function _mergeIntoStore (store, newMap, client) {
-  const updated = [];
-  const added = [];
-  const mergedStore = Object.entries(newMap).reduce((acc, [hex, aircraft]) => {
-    let logArray;
-    if (store.aircraft[hex]) {
-      logArray = updated;
-    } else {
-      logArray = added;
-      if (store.purged.has(hex)) {
-        store.purged.delete(hex);
-        logger.info({
-          message: 'add purged aircraft',
-          hex,
-          seen: aircraft.seen
-        });
-      }
-    }
-    logArray.push(hex);
-    acc.aircraft[hex] = aircraft;
-    // client.hset(store.name, hex, JSON.stringify(aircraft)).then(_ => {
-    //   console.log('set ' + hex)
-    // });
+function mapifyAircraftArray (aircraftArray = []) {
+  return aircraftArray.reduce((acc, aircraft) => {
+    acc[aircraft.hex] = aircraft;
     return acc;
-  }, _.cloneDeep(store));
-  const size = Object.keys(mergedStore.aircraft).length;
-  logger.info({
-    message: 'add new aircraft',
-    store: store.name,
-    numAdded: added.length,
-    newSize: size
-  });
-  logger.info({
-    message: 'update existing aircraft',
-    store: store.name,
-    numUpdated: updated.length,
-    size
-  });
-  return mergedStore;
+  }, {});
 }
 
 /**
@@ -202,89 +70,46 @@ function _mergeIntoStore (store, newMap, client) {
  * @returns validated and filtered serve1090 aircraft map
  * @private
  */
-function _validateAircraftMap (map) {
-  const errors = {};
-  const validatedMap = Object.entries(map).reduce((acc, [hex, aircraft]) => {
+async function validateAndWrite (client, store) {
+  client.pipeline();
+  Object.entries(store).forEach(([hex, aircraft]) => {
     const { value: validatedBody, error } = AIRCRAFT_SCHEMA.validate(aircraft);
     if (!error) {
       validatedBody.error = false;
-      acc[hex] = validatedBody;
+      client.hsetJsonEx('store:valid', hex, validatedBody, 10);
     } else {
-      errors[hex] = error.message.replace(/"/g, '\'');
+      aircraft.error = error.message.replace(/"/g, '\'');
+      client.hsetJsonEx('store:invalid', hex, aircraft, 60);
     }
-    return acc;
-  }, {});
-  return {
-    validatedMap,
-    errors
-  };
+    client.hsetJsonEx('store:all', hex, aircraft, 60);
+  });
+  await client.exec();
 }
 
 // RETRIEVE
 
-function getRawAircraft () {
-  _checkIfInitialized(this.initialized);
-  return _exportStore(this.currentRawAircraft);
+function getAllAircraft (client) {
+  return getStore(client, 'store:all');
 }
 
-function getValidAircraft () {
-  _checkIfInitialized(this.initialized);
-  return _exportStore(this.currentValidAircraft);
+async function getValidAircraft (client) {
+  return getStore(client, 'store:valid');
 }
 
-function getExcludedAircraft () {
-  _checkIfInitialized(this.initialized);
-  const currentAircraft = this.currentValidAircraft;
-  const excluded = _.pickBy(this.currentRawAircraft, a => !currentAircraft[a.hex]);
-  return _exportStore(excluded);
+async function getInvalidAircraft (client) {
+  return getStore(client, 'store:invalid');
 }
 
-function _exportStore (store) {
-  try {
-    const aircraft = Object.values(store.aircraft) || [];
-    return {
-      now: Date.now(),
-      numAircraft: aircraft.length,
-      aircraft
-    };
-  } catch (e) {
-    return {
-      now: Date.now(),
-      error: e.message
-    };
-  }
-}
-
-// UTILS
-
-function _checkIfInitialized (initialized) {
-  if (!initialized) throw new StoreError('store not initialized');
-}
-
-/**
- * Init a new aircraft store
- */
-function _createStore (name) {
+async function getStore (client, store) {
+  const aircraft = await client.hgetAllJsonValues(store);
   return {
-    name,
-    aircraft: {},
-    purged: new Set()
+    now: Date.now(),
+    count: aircraft.length,
+    aircraft
   };
 }
 
-function _setUpdated (aircraft) {
+function setUpdated (aircraft) {
   aircraft.updated = Date.now();
   return aircraft;
-}
-
-/**
- *
- * @param aircraftArray array of aircraft from dump1090
- * @returns { hex: aircraft } hex mapped to aircraft object
- */
-function _mapifyAircraftArray (aircraftArray = []) {
-  return aircraftArray.reduce((acc, aircraft) => {
-    acc[aircraft.hex] = aircraft;
-    return acc;
-  }, {});
 }
